Make multithreading named consistantly
