+++ web/lib/src/multithreaded/multithreaded-pool.worker.ts
@@ -1,9 +1,7 @@
 import * as maplibre from "../wasm/maplibre"
 
 onmessage = async message => {
-    const memory = new WebAssembly.Memory({initial: 1024, shared: false})
-    let module = message.data[0];
-    const initialised = maplibre.default(module, memory).catch(err => {
+    const initialised = maplibre.default(message.data[0], message.data[1]).catch(err => {
         // Propagate to main `onerror`:
         setTimeout(() => {
             throw err;
@@ -15,9 +13,7 @@ onmessage = async message => {
     self.onmessage = async message => {
         // This will queue further commands up until the module is fully initialised:
         await initialised;
-        let procedure_ptr = message.data[0];
-        let input = message.data[1];
-        // @ts-ignore TODO
-        await maplibre.unsync_worker_entry(procedure_ptr, input);
+        // @ts-ignore TODO may not exist
+        await maplibre.multithreaded_worker_entry(message.data);
     };
 }
