+++ web/src/platform/mod.rs
@@ -4,7 +4,7 @@ use maplibre::error::Error;
 pub mod http_client;
 
 #[cfg(target_feature = "atomics")]
-pub mod sync;
+pub mod multithreaded;
 
 #[cfg(not(target_feature = "atomics"))]
-pub mod unsync;
+pub mod singlethreaded;
