+++ web/src/platform/multithreaded/pool.rs
@@ -146,7 +146,7 @@ impl PoolState {
 
 /// Entry point invoked by the worker.
 #[wasm_bindgen]
-pub async fn sync_worker_entry(ptr: u32) -> Result<(), JsValue> {
+pub async fn multithreaded_worker_entry(ptr: u32) -> Result<(), JsValue> {
     let ptr = unsafe { Box::from_raw(ptr as *mut Work) };
     JsFuture::from((ptr.func)()).await?;
     Ok(())
